refactor(products): keep initNewFeature in copyProductForOrgs and import it
