+++ server/src/internal/products/handlers/handleCopyProduct/copyMissingFeatures.ts
@@ -1,32 +1,9 @@
-import {
-	type AppEnv,
-	type CreateFeature,
-	CreateFeatureSchema,
-	ErrCode,
-	type Feature,
-} from "@autumn/shared";
+import { CreateFeatureSchema, ErrCode } from "@autumn/shared";
 import type { AutumnContext } from "@/honoUtils/HonoEnv.js";
 import { FeatureService } from "@/internal/features/FeatureService.js";
 import type { PlanCopySource } from "@/internal/products/productUtils.js";
 import RecaseError from "@/utils/errorUtils.js";
-import { generateId } from "@/utils/genUtils.js";
-
-const initNewFeature = ({
-	data,
-	orgId,
-	env,
-}: {
-	data: CreateFeature;
-	orgId: string;
-	env: AppEnv;
-}): Feature => ({
-	...data,
-	org_id: orgId,
-	env,
-	created_at: Date.now(),
-	internal_id: generateId("fe"),
-	archived: false,
-});
+import { initNewFeature } from "./copyProductForOrgs.js";
 
 /**
  * Inserts the selected source features the target lacks, appending them to
